Move `contains_eval`
